LayoutTests/fast/forms/ios/repeatedly-focus-offscreen-select-expected.txt

This test verifies that focusing an offscreen select element does not present an offscreen popover on iPad. To manually run the test, tap the button three times.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS clickCount became 3
PASS document.activeElement is select
PASS successfullyParsed is true

TEST COMPLETE
Focus the offscreen select
